refactor: split model/constraints into package (types/facts/registry/capabilities/serialization)
