+++ datamimic_ce/model/constraints/facts.py
@@ -20,11 +20,6 @@ The JSON schema exposure is derived by a shared callable factory
 This ensures one declaration, all consumers derive consistently.
 """
 
-from collections.abc import Callable
-from dataclasses import dataclass
-from enum import StrEnum
-from typing import Any
-
 # Attribute constants (imported at module level to avoid circular imports)
 from datamimic_ce.constants.attribute_constants import (
     ATTR_CONDITION,
@@ -64,429 +59,25 @@ from datamimic_ce.constants.data_type_constants import (
     DATA_TYPE_BINARY,
     DATA_TYPE_BOOL,
     DATA_TYPE_DECIMAL,
-    DATA_TYPE_DICT,
     DATA_TYPE_FLOAT,
     DATA_TYPE_INT,
     DATA_TYPE_LIST,
     DATA_TYPE_LITERAL,
     DATA_TYPE_STRING,
 )
-from datamimic_ce.constants.element_constants import (
-    EL_ARRAY,
-    EL_ASSERT,
-    EL_CONDITION,
-    EL_DATABASE,
-    EL_DEMOGRAPHICS,
-    EL_ECHO,
-    EL_ELEMENT,
-    EL_ELSE,
-    EL_ELSE_IF,
-    EL_EXECUTE,
-    EL_FIELD,
-    EL_GENERATE,
-    EL_GENERATOR,
-    EL_ID,
-    EL_IF,
-    EL_INCLUDE,
-    EL_ITEM,
-    EL_ITERATE,
-    EL_KEY,
-    EL_LIST,
-    EL_MEMSTORE,
-    EL_MONGODB,
-    EL_NESTED_KEY,
-    EL_REFERENCE,
-    EL_SETUP,
-    EL_STATE_MACHINE,
-    EL_TRANSITION,
-    EL_VALUE,
-    EL_VARIABLE,
-    EL_WHILE,
+from datamimic_ce.model.constraints.types import (
+    AllOrNone,
+    AllowedValuesWhen,
+    Constraint,
+    Forbids,
+    ForbidsWhenValue,
+    MutuallyExclusive,
+    MutuallyExclusiveWhen,
+    RequiredOneOf,
+    Requires,
+    RequiresWhenValue,
+    ValidValues,
 )
-
-
-class DynamicSourceKind(StrEnum):
-    PYTHON_EXPRESSION = "python_expression"
-    BRACED_EXPRESSION = "braced_expression"
-
-
-class SourceFileFormat(StrEnum):
-    """Canonical runtime file formats accepted by ``source=`` consumers."""
-
-    DBUNIT_XML = ".dbunit.xml"
-    WEIGHTED_ENTITY_CSV = ".wgt.ent.csv"
-    WEIGHTED_CSV = ".wgt.csv"
-    CSV = ".csv"
-    JSON = ".json"
-    XLSX = ".xlsx"
-    XML = ".xml"
-    FIXED_WIDTH = ".fcw"
-
-
-@dataclass(frozen=True)
-class SourceCapability:
-    """One runtime-supported ``source=`` context.
-
-    ``source_type`` narrows shape-sensitive consumers such as ``nestedKey``.
-    File-loader implementations stay in their runtime tasks; this fact owns only
-    which source kinds and suffixes each element is allowed to dispatch.
-    """
-
-    element: str
-    file_formats: tuple[SourceFileFormat, ...] = ()
-    source_type: str | None = None
-    allows_memstore: bool = False
-    allows_client: bool = False
-    dynamic_source: DynamicSourceKind | None = None
-
-
-_SOURCE_CAPABILITIES: tuple[SourceCapability, ...] = (
-    SourceCapability(
-        EL_GENERATE,
-        (
-            SourceFileFormat.DBUNIT_XML,
-            SourceFileFormat.CSV,
-            SourceFileFormat.JSON,
-            SourceFileFormat.XLSX,
-            SourceFileFormat.XML,
-            SourceFileFormat.FIXED_WIDTH,
-        ),
-        allows_memstore=True,
-        allows_client=True,
-    ),
-    SourceCapability(
-        EL_ITERATE,
-        (
-            SourceFileFormat.DBUNIT_XML,
-            SourceFileFormat.CSV,
-            SourceFileFormat.JSON,
-            SourceFileFormat.XLSX,
-            SourceFileFormat.XML,
-            SourceFileFormat.FIXED_WIDTH,
-        ),
-        allows_memstore=True,
-        allows_client=True,
-    ),
-    SourceCapability(
-        EL_VARIABLE,
-        (
-            SourceFileFormat.WEIGHTED_ENTITY_CSV,
-            SourceFileFormat.CSV,
-            SourceFileFormat.JSON,
-            SourceFileFormat.XLSX,
-            SourceFileFormat.FIXED_WIDTH,
-        ),
-        allows_memstore=True,
-        allows_client=True,
-        dynamic_source=DynamicSourceKind.PYTHON_EXPRESSION,
-    ),
-    SourceCapability(
-        EL_NESTED_KEY,
-        allows_memstore=True,
-        dynamic_source=DynamicSourceKind.BRACED_EXPRESSION,
-    ),
-    SourceCapability(
-        EL_NESTED_KEY,
-        (SourceFileFormat.CSV, SourceFileFormat.JSON),
-        source_type=DATA_TYPE_LIST,
-    ),
-    SourceCapability(EL_NESTED_KEY, (SourceFileFormat.JSON,), source_type=DATA_TYPE_DICT),
-    SourceCapability(EL_KEY, (SourceFileFormat.WEIGHTED_CSV,)),
-    SourceCapability(EL_ID, (SourceFileFormat.WEIGHTED_CSV,)),
-    SourceCapability(EL_ELEMENT, (SourceFileFormat.WEIGHTED_CSV,)),
-    SourceCapability(EL_REFERENCE, allows_client=True),
-)
-
-
-def source_capabilities() -> tuple[SourceCapability, ...]:
-    """Return the immutable runtime source-capability catalog."""
-
-    return _SOURCE_CAPABILITIES
-
-
-def serialize_source_capability(capability: SourceCapability) -> dict[str, object]:
-    """Project one source-capability fact to reference/capabilities JSON."""
-
-    return {
-        "element": capability.element,
-        "source_type": capability.source_type,
-        "file_suffixes": [file_format.value for file_format in capability.file_formats],
-        "allows_memstore": capability.allows_memstore,
-        "allows_client": capability.allows_client,
-        "dynamic_source": capability.dynamic_source.value if capability.dynamic_source is not None else None,
-    }
-
-
-def _source_capabilities_for(element: str, source_type: str | None = None) -> tuple[SourceCapability, ...]:
-    return tuple(
-        capability
-        for capability in _SOURCE_CAPABILITIES
-        if capability.element == element and capability.source_type in (None, source_type)
-    )
-
-
-def supported_source_file_formats(
-    element: str,
-    source_type: str | None = None,
-) -> tuple[SourceFileFormat, ...]:
-    """File formats accepted by one element/shape, longest suffix first."""
-
-    formats = {
-        file_format
-        for capability in _source_capabilities_for(element, source_type)
-        for file_format in capability.file_formats
-    }
-    return tuple(sorted(formats, key=lambda file_format: (-len(file_format.value), file_format.value)))
-
-
-def recognized_source_file_formats() -> tuple[SourceFileFormat, ...]:
-    """All file formats known anywhere in CE, for unsupported-context diagnostics."""
-
-    formats = {file_format for capability in _SOURCE_CAPABILITIES for file_format in capability.file_formats}
-    return tuple(sorted(formats, key=lambda file_format: (-len(file_format.value), file_format.value)))
-
-
-def source_file_format(source: str) -> SourceFileFormat | None:
-    """Return a file format recognized by any runtime source context, if present."""
-
-    return next(
-        (file_format for file_format in recognized_source_file_formats() if source.endswith(file_format.value)),
-        None,
-    )
-
-
-def source_file_format_for(
-    element: str,
-    source: str,
-    source_type: str | None = None,
-) -> SourceFileFormat | None:
-    """Return the typed format this concrete runtime source context dispatches."""
-
-    return next(
-        (
-            file_format
-            for file_format in supported_source_file_formats(element, source_type)
-            if source.endswith(file_format.value)
-        ),
-        None,
-    )
-
-
-def is_source_file(source: str, element: str, source_type: str | None = None) -> bool:
-    """Whether one concrete runtime source context supports ``source`` as a file."""
-
-    return source_file_format_for(element, source, source_type) is not None
-
-
-def source_allows_memstore(element: str, source_type: str | None = None) -> bool:
-    return any(capability.allows_memstore for capability in _source_capabilities_for(element, source_type))
-
-
-def source_allows_client(element: str, source_type: str | None = None) -> bool:
-    return any(capability.allows_client for capability in _source_capabilities_for(element, source_type))
-
-
-def source_dynamic_kind(element: str, source_type: str | None = None) -> DynamicSourceKind | None:
-    return next(
-        (
-            capability.dynamic_source
-            for capability in _source_capabilities_for(element, source_type)
-            if capability.dynamic_source is not None
-        ),
-        None,
-    )
-
-
-@dataclass(frozen=True)
-class RequiredOneOf:
-    """At least one of these XML attributes must be present.
-
-    Example: a <key> must define one of type/source/values/script/generator/constant/pattern/string.
-    """
-
-    attrs: frozenset[str]
-    message: str | None = None
-    lint_only: bool = False
-
-
-@dataclass(frozen=True)
-class MutuallyExclusive:
-    """At most one of these attributes may be present.
-
-    Design note: count vs minCount/maxCount is XOR-decomposed as TWO pairwise constraints
-    (count ⊕ minCount, count ⊕ maxCount) rather than a single 3-member set, to permit
-    the legal minCount+maxCount combination.
-    """
-
-    attrs: frozenset[str]
-    message: str | None = None
-    lint_only: bool = False
-
-
-@dataclass(frozen=True)
-class MutuallyExclusiveWhen:
-    """At most one of ``attrs`` may be present when ``when_attr`` is set/truthy.
-
-    Example: with ``source=`` present, ``type=`` and ``selector=`` are alternative
-    source entity/query selectors and must not be combined.
-    """
-
-    when_attr: str
-    attrs: frozenset[str]
-    when_true: bool = False
-    message: str | None = None
-    lint_only: bool = False
-
-
-@dataclass(frozen=True)
-class Requires:
-    """If attr is present (or truthy when when_true=True), at least one of needs must be present.
-
-    Design note: when_true=True gates on _attr_true() to match the engine's existing
-    unique/cyclic checks, which permit <key unique="false" weights=...> (presence-based
-    checks would wrongly reject this as a violation). Use when_true=True only for
-    boolean-valued attributes that have legitimate false states.
-    """
-
-    attr: str
-    needs: frozenset[str]
-    when_true: bool = False
-    message: str | None = None
-    lint_only: bool = False
-
-
-@dataclass(frozen=True)
-class RequiresWhenValue:
-    """When ``when_attr`` has one of ``when_values``, require one of ``needs``.
-
-    ``unless`` provides explicit escape attributes. This captures mode rules such
-    as ``nestedKey type=list`` needing a count unless a source supplies the size.
-    """
-
-    when_attr: str
-    when_values: frozenset[str]
-    needs: frozenset[str]
-    unless: frozenset[str] = frozenset()
-    message: str | None = None
-    lint_only: bool = False
-
-
-@dataclass(frozen=True)
-class AllOrNone:
-    """Either all of these attributes are present, or none are.
-
-    Example: <key> with all-or-none {min, max, granularity} for numeric ranges
-    (though technically max alone is valid for an upper bound; this captures tighter
-    invariants where they must be a group).
-    """
-
-    attrs: frozenset[str]
-    message: str | None = None
-    lint_only: bool = False
-
-
-@dataclass(frozen=True)
-class Forbids:
-    """If attr is present (or truthy when when_true=True), none of excludes may be present.
-
-    Design note: same when_true semantics as Requires, gated on _attr_true().
-    When excludes_when_true=True, an excluded attr must ALSO be truthy for violation.
-    Example: unique forbids cyclic when BOTH are truthy (not just presence-based).
-    """
-
-    attr: str
-    excludes: frozenset[str]
-    when_true: bool = False
-    excludes_when_true: bool = False
-    message: str | None = None
-    lint_only: bool = False
-
-
-@dataclass(frozen=True)
-class ForbidsWhenValue:
-    """When ``when_attr`` has one of ``when_values``, forbid ``excludes``."""
-
-    when_attr: str
-    when_values: frozenset[str]
-    excludes: frozenset[str]
-    message: str | None = None
-    lint_only: bool = False
-
-
-@dataclass(frozen=True)
-class ValidValues:
-    """The value of attr must be in the set of valid values.
-
-    The values set may be:
-    - A frozenset[str] or tuple[str, ...] (static)
-    - A zero-arg Callable[[], set[str]] (lazy, evaluated at check time)
-
-    Lazy callables enable drift-proof enum/registry-derived sources: the fact
-    references the registry supplier, not a snapshot of its contents. When the
-    registry changes, all consumers (lint, parse, agents) re-evaluate and see
-    the new set.
-
-    Design note: this constraint only validates when attr IS present. If attr
-    is absent (None), validation is skipped — this preserves optional attributes.
-    """
-
-    attr: str
-    values: frozenset[str] | tuple[str, ...] | Callable[[], set[str]]
-    message: str | None = None
-    lint_only: bool = False
-
-
-@dataclass(frozen=True)
-class AllowedValuesWhen:
-    """When when_attr is present (or truthy when when_true=True), attr if PRESENT
-    must be in the set of allowed values.
-
-    Design note: Similar to ValidValues but gated on another attribute's presence/truthiness.
-    Presence-based check: when_attr is present -> attr must be in allowed (if attr is present).
-    Truthiness-based check: when_true=True gates on _attr_true(when_attr) -> attr must be in allowed.
-    Absent attr is always fine (the constraint only applies when both conditions are true).
-
-    Example: unique=true pins distribution to 'random' — when unique is true and distribution
-    is present, distribution must be 'random'. Missing distribution is fine (defaults to random).
-    """
-
-    attr: str
-    allowed: frozenset[str] | tuple[str, ...] | Callable[[], set[str]]
-    when_attr: str
-    when_true: bool = False
-    message: str | None = None
-    lint_only: bool = False
-
-
-# Type alias for any constraint in the vocabulary
-Constraint = (
-    RequiredOneOf
-    | MutuallyExclusive
-    | MutuallyExclusiveWhen
-    | Requires
-    | RequiresWhenValue
-    | AllOrNone
-    | Forbids
-    | ForbidsWhenValue
-    | ValidValues
-    | AllowedValuesWhen
-)
-
-
-def resolved_values(fact: ValidValues) -> frozenset[str]:
-    """The fact's valid-value set with a lazy callable resolved — the ONE typed API for
-    consumers that need the concrete set (validators, lint rules), so the
-    static-or-callable union is narrowed here once instead of ad-hoc at every call site."""
-    values = fact.values
-    return frozenset(values() if callable(values) else values)
-
-
-def resolved_allowed(fact: AllowedValuesWhen) -> frozenset[str]:
-    """The fact's allowed-value set with a lazy callable resolved — similar to resolved_values
-    but for AllowedValuesWhen constraints."""
-    allowed = fact.allowed
-    return frozenset(allowed() if callable(allowed) else allowed)
-
 
 # ============================================================================
 # Module-level constraint constants used by multiple models and delegating
@@ -1006,198 +597,3 @@ _REFERENCE_RULES: tuple[Constraint, ...] = (
     UNIQUE_DISTRIBUTION_RANDOM,
     SOURCE_DISTRIBUTION_VALUES,
 )
-
-# Explicit entry for every CE registry tag. Empty tuples are intentional and make
-# omissions review-visible: a new built-in element must decide its rule contract.
-_ELEMENT_CONSTRAINTS: dict[str, tuple[Constraint, ...]] = {
-    EL_SETUP: (),
-    EL_GENERATE: _GENERATE_RULES,
-    EL_ITERATE: (*_GENERATE_RULES, ITERATE_REQUIRES_SOURCE),
-    EL_KEY: _KEY_RULES,
-    EL_ID: _KEY_RULES,
-    EL_ELEMENT: _KEY_RULES,
-    EL_VARIABLE: _VARIABLE_RULES,
-    EL_NESTED_KEY: _NESTED_KEY_RULES,
-    EL_ARRAY: _ARRAY_RULES,
-    EL_VALUE: (),
-    EL_LIST: (),
-    EL_ITEM: (),
-    EL_REFERENCE: _REFERENCE_RULES,
-    EL_FIELD: (),
-    EL_INCLUDE: (),
-    EL_MEMSTORE: (),
-    EL_EXECUTE: _EXECUTE_RULES,
-    EL_DATABASE: (),
-    EL_MONGODB: (),
-    EL_IF: (),
-    EL_ELSE_IF: (),
-    EL_ELSE: (),
-    EL_CONDITION: (),
-    EL_ECHO: (),
-    EL_GENERATOR: (),
-    EL_DEMOGRAPHICS: (),
-    EL_STATE_MACHINE: (),
-    EL_TRANSITION: (),
-    EL_WHILE: (),
-    EL_ASSERT: (),
-}
-
-_EXTENSION_CONSTRAINTS: dict[str, tuple[Constraint, ...]] = {}
-_RULE_REGISTRY_REVISION = 0
-
-
-def element_constraints(tag: str) -> tuple[Constraint, ...]:
-    """Return the centrally registered rule facts for one CE DSL tag."""
-    return _EXTENSION_CONSTRAINTS.get(tag, _ELEMENT_CONSTRAINTS.get(tag, ()))
-
-
-def register_element_constraints(tag: str, constraints: tuple[Constraint, ...]) -> None:
-    """Register the business-rule contract for one extension element."""
-    global _RULE_REGISTRY_REVISION
-
-    if tag in _ELEMENT_CONSTRAINTS or tag in _EXTENSION_CONSTRAINTS:
-        raise ValueError(f"business-rule contract for <{tag}> is already registered")
-    _EXTENSION_CONSTRAINTS[tag] = constraints
-    _RULE_REGISTRY_REVISION += 1
-
-
-def unregister_element_constraints(tag: str) -> None:
-    """Remove an extension rule contract; built-in contracts are immutable."""
-    global _RULE_REGISTRY_REVISION
-
-    if tag not in _EXTENSION_CONSTRAINTS:
-        raise KeyError(f"extension business-rule contract for <{tag}> is not registered")
-    del _EXTENSION_CONSTRAINTS[tag]
-    _RULE_REGISTRY_REVISION += 1
-
-
-def rule_registry_revision() -> int:
-    """Return a monotonic revision used by authoring projections."""
-    return _RULE_REGISTRY_REVISION
-
-
-def registered_rule_tags() -> frozenset[str]:
-    """Tags with an explicit rule-registry decision, including intentional empties."""
-    return frozenset((*_ELEMENT_CONSTRAINTS, *_EXTENSION_CONSTRAINTS))
-
-
-def serialize_constraints(constraints: tuple[Constraint, ...]) -> list[dict[str, Any]]:
-    """Serialize a tuple of constraints to a list of dicts with kind discriminators.
-
-    Used by:
-    - constraints_schema_extra: inject into Pydantic model JSON schemas
-    - capabilities_manifest: expose constraints in the DSL reference
-    - element_reference: render prose descriptions
-
-    Each constraint becomes a dict with a "kind" discriminator and fact-specific fields.
-
-    Args:
-        constraints: A tuple of Constraint instances (from a model's __constraints__)
-
-    Returns:
-        A list of serialized constraint dicts, empty if constraints is empty
-    """
-    if not constraints:
-        return []
-
-    serialized: list[dict[str, Any]] = []
-    for fact in constraints:
-        serialized_fact: dict[str, Any] = {"kind": _constraint_kind(fact)}
-        serialized_fact.update(_serialized_constraint_fields(fact))
-
-        # Only serialize lint_only if True (non-default)
-        if fact.lint_only:
-            serialized_fact["lint_only"] = True
-
-        # Only serialize message if present (non-default)
-        if fact.message is not None:
-            serialized_fact["message"] = fact.message
-
-        serialized.append(serialized_fact)
-
-    return serialized
-
-
-def _serialized_constraint_fields(fact: Constraint) -> dict[str, Any]:
-    if isinstance(fact, RequiredOneOf | MutuallyExclusive | AllOrNone):
-        return {"attrs": sorted(fact.attrs)}
-    if isinstance(fact, MutuallyExclusiveWhen):
-        return {
-            "when_attr": fact.when_attr,
-            "attrs": sorted(fact.attrs),
-            "when_true": fact.when_true,
-        }
-    if isinstance(fact, Requires):
-        return {
-            "attr": fact.attr,
-            "needs": sorted(fact.needs),
-            "when_true": fact.when_true,
-        }
-    if isinstance(fact, RequiresWhenValue):
-        return {
-            "when_attr": fact.when_attr,
-            "when_values": sorted(fact.when_values),
-            "needs": sorted(fact.needs),
-            "unless": sorted(fact.unless),
-        }
-    if isinstance(fact, Forbids):
-        return {
-            "attr": fact.attr,
-            "excludes": sorted(fact.excludes),
-            "when_true": fact.when_true,
-            "excludes_when_true": fact.excludes_when_true,
-        }
-    if isinstance(fact, ForbidsWhenValue):
-        return {
-            "when_attr": fact.when_attr,
-            "when_values": sorted(fact.when_values),
-            "excludes": sorted(fact.excludes),
-        }
-    if isinstance(fact, ValidValues):
-        return {"attr": fact.attr, "values": sorted(resolved_values(fact))}
-    if isinstance(fact, AllowedValuesWhen):
-        return {
-            "attr": fact.attr,
-            "allowed": sorted(resolved_allowed(fact)),
-            "when_attr": fact.when_attr,
-            "when_true": fact.when_true,
-        }
-    raise TypeError(f"Unknown constraint type: {type(fact)}")
-
-
-def constraints_schema_extra(
-    constraints: tuple[Constraint, ...],
-) -> Callable[[dict[str, Any]], None]:
-    """Bind one explicit rule tuple to a Pydantic JSON-schema projector."""
-
-    def inject(schema: dict[str, Any]) -> None:
-        if constraints:
-            schema["constraints"] = serialize_constraints(constraints)
-
-    return inject
-
-
-def _constraint_kind(fact: Constraint) -> str:
-    """Map a constraint instance to its kind discriminator string."""
-    if isinstance(fact, RequiredOneOf):
-        return "required_one_of"
-    elif isinstance(fact, MutuallyExclusive):
-        return "mutually_exclusive"
-    elif isinstance(fact, MutuallyExclusiveWhen):
-        return "mutually_exclusive_when"
-    elif isinstance(fact, Requires):
-        return "requires"
-    elif isinstance(fact, RequiresWhenValue):
-        return "requires_when_value"
-    elif isinstance(fact, AllOrNone):
-        return "all_or_none"
-    elif isinstance(fact, Forbids):
-        return "forbids"
-    elif isinstance(fact, ForbidsWhenValue):
-        return "forbids_when_value"
-    elif isinstance(fact, ValidValues):
-        return "valid_values"
-    elif isinstance(fact, AllowedValuesWhen):
-        return "allowed_values_when"
-    else:
-        raise TypeError(f"Unknown constraint type: {type(fact)}")
